+++ src/extract_crate_docs/resolver/index/parents.rs
@@ -6,25 +6,43 @@
 
 use std::collections::{HashMap, hash_map::Entry};
 
+use color_eyre::eyre::{Result, bail};
 use rustdoc_types::Id;
+use tracing::error_span;
 
 use super::simple::{SimpleItem, SimpleItemKind};
 
-pub fn parents(index: &HashMap<Id, SimpleItem>, root: Id) -> HashMap<Id, Id> {
+const RECURSION_LIMIT: usize = 64;
+
+pub fn parents(index: &HashMap<Id, SimpleItem>, root: Id) -> Result<HashMap<Id, Id>> {
     let mut parents = HashMap::new();
-    parents_recurse(index, &mut parents, root, 0);
-    parents.into_iter().map(|(child_id, parent)| (child_id, parent.id)).collect()
+    parents_recurse(index, &mut parents, root, 0, &mut vec![])?;
+    Ok(parents.into_iter().map(|(child_id, parent)| (child_id, parent.id)).collect())
 }
 
-fn parents_recurse(
-    index: &HashMap<Id, SimpleItem>,
+fn parents_recurse<'a>(
+    index: &HashMap<Id, SimpleItem<'a>>,
     parents: &mut HashMap<Id, Parent>,
     parent_id: Id,
     depth: usize,
-) {
+    path_for_error: &mut Vec<&'a str>,
+) -> Result<()> {
+    if path_for_error.len() > RECURSION_LIMIT {
+        let item_path = path_for_error
+            .iter()
+            .filter_map(|&s| (!s.is_empty()).then_some(s))
+            .collect::<Vec<_>>()
+            .join("::");
+
+        let _span = error_span!("", %item_path).entered();
+        bail!("recursed too deep while resolving item paths ({RECURSION_LIMIT})");
+    }
+
     let Some(parent_item) = index.get(&parent_id) else {
-        return;
+        return Ok(());
     };
+
+    path_for_error.push(parent_item.name);
 
     let parent_is_use = matches!(&parent_item.kind, SimpleItemKind::Use { .. });
 
@@ -59,8 +77,11 @@ fn parents_recurse(
         // A `use` does not count towards depth.
         let child_depth = if parent_is_use || child_is_use { depth } else { depth + 1 };
 
-        parents_recurse(index, parents, child_id, child_depth);
+        parents_recurse(index, parents, child_id, child_depth, path_for_error)?;
+        path_for_error.pop();
     }
+
+    Ok(())
 }
 
 #[derive(Clone, Copy)]